Add follow and unfollow activity. Add array index check. Fix getProfilePageOfOtherUser typo. Add followManager.
