+++ Project/Wikode/templates/wikodeApp/homePage.html
@@ -94,21 +94,21 @@
                       <li class="list-group-item">
                         <div class="media-body">
                           <p>
-                              <a href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 3 %}">Kenan Kural</a> - <a>14 followers</a>
+                              <a href="{% url 'wikodeApp:getProfilePageOfOtherUser' 3 %}">Kenan Kural</a> - <a>14 followers</a>
                           </p>
                         </div>
                       </li>
                       <li class="list-group-item">
                         <div class="media-body">
                           <p>
-                              <a href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 2 %}">Can Dayan</a> - <a>8 followers</a>
+                              <a href="{% url 'wikodeApp:getProfilePageOfOtherUser' 2 %}">Can Dayan</a> - <a>8 followers</a>
                           </p>
                         </div>
                       </li>
                       <li class="list-group-item">
                         <div class="media-body">
                           <p>
-                              <a href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 5 %}">Emre Çalık</a> - <a>9 followers</a>
+                              <a href="{% url 'wikodeApp:getProfilePageOfOtherUser' 5 %}">Emre Çalık</a> - <a>9 followers</a>
                           </p>
                         </div>
                       </li>
@@ -149,8 +149,8 @@
                         <div class="media-body">
                           <h5 class="media-heading">
                             <small class="pull-right">2h ago</small>
-                            <a class="name" href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 3 %}">Kenan Kural</a> started following
-                            <a href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 5 %}">Emre Çalık</a>.
+                            <a class="name" href="{% url 'wikodeApp:getProfilePageOfOtherUser' 3 %}">Kenan Kural</a> started following
+                            <a href="{% url 'wikodeApp:getProfilePageOfOtherUser' 5 %}">Emre Çalık</a>.
                           </h5>
                           <small>Today 2:55 pm - 31.10.2021</small>
                         </div>
@@ -161,7 +161,7 @@
                         <div class="media-body">
                           <h5 class="media-heading">
                             <small class="pull-right">6h ago</small>
-                            <a href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 2 %}">Can Dayan</a> upvoted tag
+                            <a href="{% url 'wikodeApp:getProfilePageOfOtherUser' 2 %}">Can Dayan</a> upvoted tag
                             <a class="name">SARS-CoV-2</a>.
                           </h5>
                           <small>Today 12:00 am - 31.10.2021</small>
@@ -173,8 +173,8 @@
                         <div class="media-body">
                           <h5 class="media-heading">
                             <small class="pull-right">20h ago</small>
-                            <a href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 2 %}">Can Dayan</a> started following
-                            <a href="{% url 'wikodeApp:getProfilePageOfOtherfUser' 5 %}">Emre Çalık</a>.
+                            <a href="{% url 'wikodeApp:getProfilePageOfOtherUser' 2 %}">Can Dayan</a> started following
+                            <a href="{% url 'wikodeApp:getProfilePageOfOtherUser' 5 %}">Emre Çalık</a>.
                           </h5>
                           <small>Yesterday 9:00 pm - 30.10.2021</small>
                         </div>
