+++ Project/Wikode/wikodeApp/urls.py
@@ -9,7 +9,7 @@ urlpatterns = [
     path('registrationRequests/', views.registrationRequests, name='registrationRequests'),
     path('userLogin/', views.userLogin, name='userLogin'),
     path('profile/', views.myProfilePage, name='myProfilePage'),
-    path('profile/<int:pk>', views.getProfilePageOfOtherfUser, name='getProfilePageOfOtherfUser'),
+    path('profile/<int:pk>', views.getProfilePageOfOtherUser, name='getProfilePageOfOtherUser'),
     path('userList/', views.userList, name='userList'),
     path('getArticles/', views.getArticles, name='getArticles'),
     path('articleDetail/<int:pk>', views.articleDetail, name='articleDetail'),
